Move formatMessages to ai
